Fix runtime error in jpv3 cli
