+++ .gitignore
@@ -34,3 +34,4 @@ jiiify-presentation-v3.iml
 
 # Output of native builds
 output.zip
+output.csv.zip
