+++ src/main/resources/META-INF/services/java.nio.file.spi.FileTypeDetector
@@ -1,2 +1,2 @@
-info.freelibrary.iiif.presentation.v3.utils.cmdline.JPv3FileDetector
+info.freelibrary.iiif.presentation.v3.utils.cmdline.JPv3Utils$JPv3FileDetector
